rename object package to maps
