+++ support/maps/maps.go
@@ -1,4 +1,4 @@
-package object
+package maps
 
 import (
 	"reflect"
