+++ support/maps/maps_test.go
@@ -1,4 +1,4 @@
-package object
+package maps
 
 import (
 	"testing"
